Add thought on men learning to socialize
I think we bear responsibility to correct
behavior of those around us.
If something is creepy to us, it is implicitly
wrong. How do we handle it? Do we label the
men that carry immense potential and help dig them
a pit, or do we call upon them to be improve?